* Separated Parser Tests into different classes. * Added toString() to ASTs
