+++ parser/src/test/kotlin/ParserMethodResultTest.kt
@@ -1,181 +1,11 @@
-import ast.AssignmentStatement
-import ast.Call
-import ast.DeclarationStatement
-import ast.LiteralArgument
-import ast.MethodResult
-import ast.Range
-import ast.Scope
-import ast.VariableArgument
-import ast.VariableDeclaration
+import ast.*
 import lexer.LexerImpl
-import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import parser.MyParser
 import parser.Parser
-import parser.ParserCommons
 
-class ParserTest {
-    @Test
-    fun test001_testSimpleLiteralVariableDeclarationnumber() {
-        val code = "let a: number = 5;" // range 14, including spaces...
-        val tokens = LexerImpl().tokenize(code)
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        val expected =
-            Result.success(
-                Scope(
-                    "program",
-                    Range(0, 17),
-                    listOf(
-                        VariableDeclaration(Range(4, 4), "a", "number", LiteralArgument(Range(16, 16), "5", "number")),
-                    ),
-                ),
-            )
-
-        assertEquals(expected, ast)
-    }
-
-    @Test
-    fun test002_testSimpleLiteralVariableDeclarationBool() {
-        val code = "let a: boolean = true;"
-        val tokens = LexerImpl().tokenize(code)
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        val expected =
-            Result.success(
-                Scope(
-                    "program",
-                    Range(0, 21),
-                    listOf(
-                        VariableDeclaration(Range(4, 4), "a", "boolean", LiteralArgument(Range(17, 20), "true", "boolean")),
-                    ),
-                ),
-            )
-
-        assertEquals(expected, ast)
-    }
-
-    @Test
-    fun test003_testSimpleLiteralVariableDeclarationNumber() {
-        val code = "let a: number = 3.5;"
-        val tokens = LexerImpl().tokenize(code)
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        val expected =
-            Result.success(
-                Scope(
-                    "program",
-                    Range(0, 19),
-                    listOf(
-                        VariableDeclaration(Range(4, 4), "a", "number", LiteralArgument(Range(16, 18), "3.5", "number")),
-                    ),
-                ),
-            )
-
-        assertEquals(expected, ast)
-    }
-
-    @Test
-    fun test004_testSimpleLiteralVariableDeclarationStringOfOneWord() {
-        val code = "let a: string = \"Juan\";"
-        val tokens = LexerImpl().tokenize(code)
-
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        val expected =
-            Result.success(
-                Scope(
-                    "program",
-                    Range(0, 22),
-                    listOf(
-                        VariableDeclaration(
-                            Range(4, 4),
-                            "a",
-                            "string",
-                            LiteralArgument(Range(16, 21), "Juan", "string"),
-                        ),
-                    ),
-                ),
-            )
-
-        assertEquals(expected, ast)
-    }
-
-    @Test
-    fun test005_testConsecutiveLiteralDeclarations() {
-        val code = "let a: string = \"Juan\"; let b: number = 5; let c: boolean = true;"
-        val tokens = LexerImpl().tokenize(code)
-
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        val expected =
-            Result.success(
-                Scope(
-                    "program",
-                    Range(0, 64),
-                    listOf(
-                        VariableDeclaration(
-                            Range(4, 4),
-                            "a",
-                            "string",
-                            LiteralArgument(Range(16, 21), "Juan", "string"),
-                        ),
-                        VariableDeclaration(Range(28, 28), "b", "number", LiteralArgument(Range(40, 40), "5", "number")),
-                        VariableDeclaration(
-                            Range(47, 47),
-                            "c",
-                            "boolean",
-                            LiteralArgument(Range(60, 63), "true", "boolean"),
-                        ),
-                    ),
-                ),
-            )
-
-        assertEquals(expected, ast)
-    }
-
-    @Test
-    fun test006_testDeclarationWithVariableArgument() {
-        val code = "let a: number = b;"
-        val tokens = LexerImpl().tokenize(code)
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        val expected =
-            Result.success(
-                Scope(
-                    "program",
-                    Range(0, 17), // Adjusted range
-                    listOf(
-                        VariableDeclaration(Range(4, 4), "a", "number", VariableArgument(Range(16, 16), "b")),
-                    ),
-                ),
-            )
-
-        assertEquals(expected, ast)
-    }
-
-    // see, it doesn't check if the type of the variable is correct
-    // but I don't think that is something for the parser to check
-    @Test
-    fun test007_testDeclarationsWithLiteralAndVariableArguments() {
-        val code = "let a: number = 5; let b: string = a;"
-        val tokens = LexerImpl().tokenize(code)
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        val expected =
-            Result.success(
-                Scope(
-                    "program",
-                    Range(0, 36),
-                    listOf(
-                        VariableDeclaration(Range(4, 4), "a", "number", LiteralArgument(Range(16, 16), "5", "number")),
-                        VariableDeclaration(Range(23, 23), "b", "string", VariableArgument(Range(35, 35), "a")),
-                    ),
-                ),
-            )
-
-        assertEquals(expected, ast)
-    }
+class ParserMethodResultTest {
 
     @Test
     fun test008_testSumMethodDeclaration() {
@@ -209,7 +39,7 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -244,7 +74,7 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -289,7 +119,7 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -333,7 +163,7 @@ class ParserTest {
                     ),
                 ),
             )
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -397,18 +227,9 @@ class ParserTest {
                     ),
                 ),
             )
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
-    @Test
-    fun test013_testSearchForClosingCharacter() {
-        val code = "((()))))"
-        val tokens = LexerImpl().tokenize(code)
-        val commons = ParserCommons()
-        val gotChar = commons.searchForClosingCharacter(tokens, "(", 0)
-        val expected = Result.success(5)
-        assertEquals(expected, gotChar)
-    }
 
     @Test
     fun test014_testMethodDeclarationWithParentheses() {
@@ -452,7 +273,7 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -507,7 +328,7 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     // solucionar el tema de que en realidad si usa parentesis non meterlos como method result. El parentesis no es un operator.
@@ -563,7 +384,7 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     // falta resolver el tema de las comas. No son operadores, son SEPARADORES.
@@ -600,7 +421,7 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -645,7 +466,7 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -690,30 +511,9 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
-    // lo que pasa en sum(3, sum(5, 2)) es que ya de una te toma tres metodos: 3, sum(5 y 2)
-    @Test
-    fun test020_testSearchForClosingBracketsMethod() {
-        val code = "let test: number = sum(3, sum(5, 2));"
-        val tokens = LexerImpl().tokenize(code)
-        val commons = ParserCommons()
-        val gotChar = commons.searchForClosingCharacter(tokens, "(", 6)
-        val expected = Result.success(15)
-        assertEquals(expected, gotChar)
-    }
-
-    @Test
-    fun test021_testSearchForClosingBracketsMethodPt2() {
-        val code = "let test: number = sum(3, sum(5, 2));"
-
-        val tokens = LexerImpl().tokenize(code)
-        val commons = ParserCommons()
-        val gotChar = commons.searchForClosingCharacter(tokens, "(", 10)
-        val expected = Result.success(14)
-        assertEquals(expected, gotChar)
-    }
 
     @Test
     fun test022_testComplexEcuationWithParenthesesButParenthesesAreNotIncludedNumberheAST() {
@@ -767,78 +567,9 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
-    @Test
-    fun test023_expresionStatementTest() {
-        val code = "let a: number;"
-        val tokens = LexerImpl().tokenize(code)
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        val expected =
-            Result.success(
-                Scope(
-                    "program",
-                    Range(0, 13),
-                    listOf(DeclarationStatement(Range(4, 4), "a", "number")),
-                ),
-            )
-        assertEquals(ast, expected)
-    }
-
-    @Test
-    fun test024_variableDeclarationnumberwoDifferentParts() {
-        val code = "let a: number; a = 54;"
-        val tokens = LexerImpl().tokenize(code)
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        val expected =
-            Result.success(
-                Scope(
-                    "program",
-                    Range(0, 21),
-                    listOf(
-                        DeclarationStatement(Range(4, 4), "a", "number"),
-                        AssignmentStatement(Range(15, 15), "a", LiteralArgument(Range(19, 20), "54", "number")),
-                    ),
-                ),
-            )
-
-        assertEquals(ast, expected)
-    }
-
-    @Test
-    fun test025_assignementStatement() {
-        val code = "a = 54;"
-        val tokens = LexerImpl().tokenize(code)
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        val expected =
-            Result.success(
-                Scope(
-                    "program",
-                    Range(0, 6),
-                    listOf(AssignmentStatement(Range(0, 0), "a", LiteralArgument(Range(4, 5), "54", "number"))),
-                ),
-            )
-
-        assertEquals(ast, expected)
-    }
-
-    @Test
-    fun test026_assignementStatementArgumentIsLiteralAndCorrect() {
-        val code = "a = 54;"
-        val tokens = LexerImpl().tokenize(code)
-        val parser: Parser = MyParser()
-        val ast = parser.parseTokens(tokens)
-        ast.onSuccess {
-            val first = it.body.first()
-            if (first is AssignmentStatement) {
-                assertEquals(first.value, LiteralArgument(Range(4, 5), "54", "number"))
-            }
-        }
-    }
 
     @Test
     fun test027_testSumOfTwoVariableArguments() {
@@ -871,7 +602,7 @@ class ParserTest {
                     ),
                 ),
             )
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -905,7 +636,7 @@ class ParserTest {
                     ),
                 ),
             )
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -927,7 +658,7 @@ class ParserTest {
                     ),
                 ),
             )
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -965,7 +696,7 @@ class ParserTest {
                     ),
                 ),
             )
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -977,7 +708,7 @@ class ParserTest {
 
         val expected =
             Result.success(Scope("program", Range(0, 8), listOf(MethodResult(Range(0, 7), Call(Range(0, 7), "object", listOf())))))
-        assertEquals(expected, ast)
+        Assertions.assertEquals(expected, ast)
     }
 
     @Test
@@ -1001,7 +732,7 @@ class ParserTest {
                     ),
                 ),
             )
-        assertEquals(ast, expected)
+        Assertions.assertEquals(ast, expected)
     }
 
     @Test
@@ -1025,7 +756,7 @@ class ParserTest {
                     ),
                 ),
             )
-        assertEquals(ast, expected)
+        Assertions.assertEquals(ast, expected)
     }
 
     @Test
@@ -1065,7 +796,7 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(ast, expected)
+        Assertions.assertEquals(ast, expected)
     }
 
     @Test
@@ -1105,6 +836,6 @@ class ParserTest {
                 ),
             )
 
-        assertEquals(ast, expected)
+        Assertions.assertEquals(ast, expected)
     }
 }
